Changes to be committed: modified:   app/Controllers/Api/BundlingController.php modified:   app/Controllers/Api/CategoryBundlingController.php modified:   app/Controllers/Api/CategoryController.php modified:   app/Controllers/Api/CourseController.php modified:   app/Controllers/Api/UserVideoController.php modified:   app/Controllers/Api/VideoCategoryController.php modified:   app/Controllers/Api/VideoController.php modified:   app/Controllers/Api/VoucherController.php

+++ app/Controllers/Api/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Controllers\Api;
 
 use CodeIgniter\RESTful\ResourceController;
 use App\Models\Category;
+use App\Models\Users;
 use CodeIgniter\HTTP\RequestInterface;
 use Firebase\JWT\JWT;
 
@@ -67,6 +68,14 @@ class CategoryController extends ResourceController
 
         try {
             $decoded = JWT::decode($token, $key, ['HS256']);
+            $user = new Users;
+
+            // cek role user
+            $data = $user->select('role')->where('id', $decoded->uid)->first();
+            if($data['role'] != 'admin'){
+                return $this->fail('Tidak dapat di akses selain admin', 400);
+            }
+
             $model = new Category();
             $rules = [
                 'name' => 'required',
@@ -118,6 +127,14 @@ class CategoryController extends ResourceController
 
         try {
             $decoded = JWT::decode($token, $key, ['HS256']);
+            $user = new Users;
+
+            // cek role user
+            $data = $user->select('role')->where('id', $decoded->uid)->first();
+            if($data['role'] != 'admin'){
+                return $this->fail('Tidak dapat di akses selain admin', 400);
+            }
+
             $model = new Category();
             $rules = [
                 'name' => 'required',
@@ -177,6 +194,14 @@ class CategoryController extends ResourceController
 
         try {
             $decoded = JWT::decode($token, $key, ['HS256']);
+            $user = new Users;
+
+            // cek role user
+            $data = $user->select('role')->where('id', $decoded->uid)->first();
+            if($data['role'] != 'admin'){
+                return $this->fail('Tidak dapat di akses selain admin', 400);
+            }
+
             $model = new Category();
             if ($model->find($id)) {
                 $model->delete($id);
